Restore the behavior of RingAttentionTest.test_ring_attention_compile on CUDA platform.
On CUDA platform, this test bundles memory efficient attention (ME) test
with flash attention (FA) test, and on platforms that do not suppport
FA, it will not test ME as well.

On ROCM we hope they can be tested independently, and refactored the
skip conditions, but this apparently enabled ME tests accidentally and
caused undiscovered failures on CUDA platforms.

This patch restores the old testing behavior on CUDA (non-ROCM) platform.